Reestructura de ContratoCompraVenta

+++ src/main/java/com/banquito/formalizacion/exception/ContratoYaExisteException.java
@@ -2,10 +2,10 @@ package com.banquito.formalizacion.exception;
 
 public class ContratoYaExisteException extends RuntimeException {
 
-    private final Integer idSolicitud;
+    private final Long idSolicitud;
     private final String tipoContrato;
 
-    public ContratoYaExisteException(Integer idSolicitud, String tipoContrato) {
+    public ContratoYaExisteException(Long idSolicitud, String tipoContrato) {
         super();
         this.idSolicitud = idSolicitud;
         this.tipoContrato = tipoContrato;
@@ -16,7 +16,7 @@ public class ContratoYaExisteException extends RuntimeException {
         return String.format("Ya existe un %s para la solicitud ID: %d", tipoContrato, idSolicitud);
     }
 
-    public Integer getIdSolicitud() {
+    public Long getIdSolicitud() {
         return idSolicitud;
     }
