+++ src/main/java/com/banquito/formalizacion/repository/ContratoCompraVentaRepository.java
@@ -12,21 +12,21 @@ import com.banquito.formalizacion.enums.ContratoVentaEstado;
 import com.banquito.formalizacion.model.ContratoCompraVenta;
 
 @Repository
-public interface ContratoCompraVentaRepository extends JpaRepository<ContratoCompraVenta, Integer> {
+public interface ContratoCompraVentaRepository extends JpaRepository<ContratoCompraVenta, Long> {
 
-    Optional<ContratoCompraVenta> findByIdSolicitud(Integer idSolicitud);
+    Optional<ContratoCompraVenta> findByIdSolicitud(Long idSolicitud);
     
     Optional<ContratoCompraVenta> findByNumeroContrato(String numeroContrato);
     
     List<ContratoCompraVenta> findByEstado(ContratoVentaEstado estado);
     
-    boolean existsByIdSolicitud(Integer idSolicitud);
+    boolean existsByIdSolicitud(Long idSolicitud);
     
     boolean existsByNumeroContrato(String numeroContrato);
 
     Page<ContratoCompraVenta> findByEstado(ContratoVentaEstado estado, Pageable pageable);
     
-    Page<ContratoCompraVenta> findByIdSolicitud(Integer idSolicitud, Pageable pageable);
+    Page<ContratoCompraVenta> findByIdSolicitud(Long idSolicitud, Pageable pageable);
     
     Page<ContratoCompraVenta> findByNumeroContratoContainingIgnoreCase(String numeroContrato, Pageable pageable);
 
@@ -34,11 +34,11 @@ public interface ContratoCompraVentaRepository extends JpaRepository<ContratoCom
         ContratoVentaEstado estado, String numeroContrato, Pageable pageable);
     
     Page<ContratoCompraVenta> findByEstadoAndIdSolicitud(
-        ContratoVentaEstado estado, Integer idSolicitud, Pageable pageable);
+        ContratoVentaEstado estado, Long idSolicitud, Pageable pageable);
     
     Page<ContratoCompraVenta> findByNumeroContratoContainingIgnoreCaseAndIdSolicitud(
-        String numeroContrato, Integer idSolicitud, Pageable pageable);
+        String numeroContrato, Long idSolicitud, Pageable pageable);
     
     Page<ContratoCompraVenta> findByEstadoAndNumeroContratoContainingIgnoreCaseAndIdSolicitud(
-        ContratoVentaEstado estado, String numeroContrato, Integer idSolicitud, Pageable pageable);
+        ContratoVentaEstado estado, String numeroContrato, Long idSolicitud, Pageable pageable);
 }
